+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fixed error occurring when angering shopkeepers
 * Fixed errors in outlaw missions occurring when setting up outlaws' moves.
 * Refactored level-reset logic to avoid potential data loss
+* Wide Lens effect -> Boost accuracy x1.5
+* Added Fickle Specs: +1 Attack Range, Torment effect, Sticky
 * Moves that temporarily learn moves such as Transform and Mimic will switch on/off for AI use based on settings.
 * Cyrillic letter support
 * Mission Mod: Fixed blackscreen with missions taking place in alternate segments